Quick hack to test AppId/AppCustId matching
This is to prevent accidental flashing of incompatible firmware.
If the bootloader hasn’t been configured with the (optional)
GetMetadata command enabled then the “-f” force flag will always have
to be used. Same thing if the AppId and/or AppCustId are all zeroes (as
that’s the default value which is useless)